+++ atomic/atom/smfetch/src/browser/smfetch/atom.js
@@ -228,6 +228,9 @@ export default await (() => {
         } else {
           let response = await fetch(furl, fdata);
           let result = {
+            code: 0,
+            data: null,
+            msg: "",
             status: response.status,
             statusText: response.statusText,
           };
@@ -240,13 +243,13 @@ export default await (() => {
               result.data = await response.blob();
             } else {
               let resp = await response.json();
-              if (ajax) result.data = await resp;
-              else result = resp;
-
-              if (achieve) achieve(result);
+              result = { ...result, ...resp };
             }
           } else {
-            if (fault) fault(result);
+            if (response.status == 301) {
+              let resp = await response.json();
+              window.location = resp.redirect;
+            }
           }
           return result;
         }