+++ Changelog.md
@@ -21,7 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix session not pass to onredict funcion in `reaction.js` at deskelectronjs engine. Completed on 2025-07-24
+- Fix session not pass to onredict funcion in `reaction.js` at deskelectronjs engine
+- Fix browser redirect both atom/smfetch and engine/webnodejs issue. Completed on 2025-08-05
 
 ### Security